requirements.txt
tensorflow-gpu==1.12.0
opencv-python
numpy
h5py
tqdm
patool
matplotlib
wandb
